Format AE specific analysis

Format AE specific analysis

Usage

format_ae_specific_subgroup(
  outdata,
  display = c("n", "prop"),
  digits_prop = 1,
  digits_ci = 1,
  digits_p = 3,
  digits_dur = c(1, 1),
  digits_events = c(1, 1),
  mock = FALSE
)

Arguments

outdata

An outdata object created by prepare_ae_specific().

display

A character vector of measurement to be displayed.

n: Number of subjects with AE.

prop: Proportion of subjects with AE.

total: Total columns.

dur: Average of AE duration.

events: Average number of AE per subject.

digits_prop

A numeric value of number of digits for proportion value.

digits_ci

A numeric value of number of digits for confidence interval

digits_p

A numeric value of number of digits for p-value .

digits_dur

A numeric value of number of digits for average duration of AE.

digits_events

A numeric value of number of digits for average of number of AE per subjects.

mock

Logical. Display mock table or not.

Value

A list of analysis raw datasets.